refactor: Move MaxBV methods into new ConstantFolder class
